pyproject.toml
[tool.poetry]
name = "recursive-ma"
version = "0.1.0"
description = ""
authors = ["Hessam Mehr <Hessam.Mehr@glasgow.ac.uk>"]
readme = "README.md"
packages = [{include = "recursive_ma"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.1"
pyinterval = "^1.2.0"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pylint = "^2.17.3"
pytest = "^7.3.1"
ipykernel = "^6.22.0"
seaborn = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
